Move affine scheme inclusion proposition & minor fixes
Moves the proposition that the inclusion of affine schemes into schemes
admits a left adjoint a bit forward to the exercise placeholder that
actually is supposed to prove it (there's no proof here yet anyhow).
Also rewords some things according to that change and fixes a few
unrelated typos here and there.