Bump `latex-packages` & discussion of "reasonable" classes
Putting the bump first because it is a pretty drastic change: This
update changes the look of the notes completely. According to the
non-cosmetic changes the update includes we remove some things from the
preamble and replace `\ref` with `\cref` and friends everywhere.

This commit also adds mathematical content (which I forgot to commit
beforehand, sorry!): Most importantly the definition and some basic
properties of "reasonable" classes of morphisms of schemes à la Vakil
and an elaboration on why the map from the affine line to the spectrum
of the underlying field is not universally closed.