Add subsection on graded rings & section on quasi-coherent sheaves
Both sections include some initial work but are far from finished. The
subsection on graded rings is subordinate to the (new) section on
projective schemes and the Proj-construction which doesn't contain much
else yet.

Also adds a minor amount of work in the properties of morphisms section,
including half a proof that the classes of open (resp. closed)
immersions are reasonable.

Finally changes the notation for nilradicals to a simple Nil over the
previous mathfrak N, and swaps the order in which references and
bibliography are printed.