Add section on Line Bundles
Includes the definition of the Picard group, the statement that it is
isomorphic to the first cohomology group (proof still missing), as well
as a discussion of how sections of line bundles relate to maps into
projective spaces and the definition of ample line bundles.

Also generalizes the definition of codimension to arbitrary closed sets.